backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
pymongo==4.6.1

# AI/ML
groq==0.4.1

# Environment and configuration
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dateutil==2.8.2

# Development (optional)
pytest==7.4.3
httpx==0.26.0
